Check prefix when iterating db
